#9 Elitism controlled by global constant, disabled
Decided to disable elitism for now until we can figure out how to keep the population size constant no matter what the elitism value is.

I think the trick may have to do with NEAT's species. Either way, a better understanding of how the population is managed is needed.

Perhaps the simple solution is to remember the population size (distinct from what is stored in the config, since that apparently changes) and only display that many shapes no matter how many are actually in the population