#4 Spawned in the basic mechanism for the switch
The player selection switch now spawns in all the basic parts of the switch that will indicate if a structure has been chosen by a player or not. It may need to be shifted over from the spot it is at now but it works. The code is also a little messy so clean up will be needed later.