#23 Created new method that detects a block
The method returns a block if there is one and none if there is not. In the query cppn for shape method, the previously mentioned method is called and used to determine whether there is a block that can be appended to shape. I think that the amount of parameters in the block detection method can be reduced but this is what I have right now that works