#41 Each function knows its own "fitness threshold"
Now the type_count works with values not 1000 (when any of the RANGES are a value other than 10). As for type_target, I tested it twice with two different values and different dimensions and it also seemed to work. The fitness_threshold seemed to be accurately changed depending on the command line parameters.