#50 console based control has options r and q
Similar to what I did in #31, I added the functionality of r and q into the console based way of evolving too.  On top of this, I added a try/except in to make sure the selections won't cause it to crash, so the only way to exit is by hitting q. Still needs comments and the save feature