Don't commit every file you create
@richeyme @alejmedinajr
Results files produced by the code should not be committed to the repository, even if the results were meaningful/useful. Be on the look out for this. Such files can be deleted, or you can simply uncheck them before making commits. In some cases, certain directories or file types should be added to the .gitignore file, but consult be before doing that.