Issue with pop size
NEAT-Python will sometimes create a population larger than the initially specified population size (I think this is related to having enough members of each species). This was creating problems since we only reserve enough corner spots for a certain number of genomes. The changes in these loops fix the issue, though this means that the population will sometimes contain an unevaluated member.