Minimum num blocks could cause problems
IF there is a restriction requiring a minimum number of blocks to be generated, the shape generation could become stuck in what looks like an infinite loop. It would terminate eventually, but it takes forever. So, I changed how this works. After 500 attempts the presence threshold is lowered so much that all blocks appear.

The 500 should probably be a command line parameter.