#42 Added command line parameter + changed snakes
I added a command line parameter that would essentially ignore if the snake was out of bounds in the positive y direction. I also made changes to the check out of bounds methods so that it only checks if the snake is out of bounds in the y direction exclude_y_coordinate is set to false. Still has to be tested.