Merge branch 'dev_richey'
Conflicts:
	cppn_generation.py

@richeyme Check that this merge results in the correct behavior. #42 .

However, I think this should also be refactored to avoid repetition. The code here mostly duplicates the code for regular redirected snakes. This is just a special case of that code, so the changes should go there.

Also, does this actually work? I thought the last direction in the list would be in the Z direction, not the Y direction.