Weird errors fixed
Something must be going wrong with the merges. Not sure why type_count was being checked for specifically. Also not sure why visualize was not being imported. There are also some potentially uninitialized variables in the code near the end, like stats and checkpointer. Need to keep an eye on these.